Change Progress property type from double to int

+++ Domain/DTOs/DownloadItem.cs
@@ -4,7 +4,7 @@ public record DownloadItem : SearchResult
 {
     public required string SavePath { get; init; }
     public required DownloadStatus Status { get; init; }
-    public required double Progress { get; init; }
+    public required int Progress { get; init; }
     public required long DownSpeed { get; init; }
     public required long UpSpeed { get; init; }
     public required long Eta { get; init; }
